Fix #42, 테스트 코드 추가
#42 버그는 프론트에서 api를 호출할 때 writer등의 필수적인 내용을 안보내서 발생했다

+++ src/main/java/uhs/alphabet/domain/service/BoardService.java
@@ -87,6 +87,7 @@ public class BoardService {
     @Transactional
     public void deletePost(Long id, String pw) {
         Optional<BoardEntity> boardEntityOptional = boardRepository.findById(id);
+        if (!boardEntityOptional.isPresent()) return;
         BoardEntity boardEntity = boardEntityOptional.get();
         if (boardEntity.getPw().toString().equals(pw)) boardRepository.deleteById(id);
     }
